% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-does-df-exist.R
\name{cs.doesDfExist}
\alias{cs.doesDfExist}
\title{Check whether dataframe exists in environment}
\usage{
cs.doesDfExist(df)
}
\arguments{
\item{df}{opal dataframe}
}
\value{
None. Stops function if df doesn't exist in one of more cohorts.
}
\description{
Again for internal use, this function can be used at the start of 
functions where users provide a dataframe to check that the df exists
When I've got more time combine this with "cs.doVarsExist"
}
