% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-do-vars-exist.R
\name{cs.doVarsExist}
\alias{cs.doVarsExist}
\title{Check whether variables exist in dataframe}
\usage{
cs.doVarsExist(df, vars)
}
\arguments{
\item{df}{opal dataframe}

\item{vars}{vector of variable names expected to be contained in dataframe}
}
\value{
a list with two elements: (i) TRUE/FALSE depending on whether all 
        provided variables are in the data frame, (ii) vector of variables
        not present in dataframe (NULL if first list element == TRUE)
}
\description{
Probably mostly for internal use, this function can be used at the start
of functions where users provide both vars and a dataframe to check that 
the vars exist in the dataframe provided
}
