% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-reg-tab.R
\name{cs.regTab}
\alias{cs.regTab}
\title{Simple function to extract key coefficients from regression models.}
\usage{
cs.regTab(model, type)
}
\arguments{
\item{model}{saved output from either ds.glm or ds.glmSLMA}

\item{type}{either "ipd" or "slma" depending on the type of analysis done}
}
\value{
A tibble containing the estimate with lower and upper confidence intervals
}
\description{
Simple function to extract key coefficients from regression models.
}
