% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-rm-lots.R
\name{cs.rmLots}
\alias{cs.rmLots}
\title{Remove multiple objects from ds environment}
\usage{
cs.rmLots(vars)
}
\arguments{
\item{vars}{objects that you want to remove}
}
\value{
None. Objects removed from ds environment
}
\description{
This is a very simple wrapper around ds.rm to allow you to remove more than 
one object at a time.
}
