% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-tidy-env.R
\name{cs.tidyEnv}
\alias{cs.tidyEnv}
\title{Tidy up the server environment}
\usage{
cs.tidyEnv(obj, type = "remove")
}
\arguments{
\item{obj}{objects that you want to either keep or remove}

\item{type}{either "remove" to remove the listed objects of "keep" to keep
the listed objects and remove everything else.}
}
\value{
None. Objects removed from ds environment
}
\description{
This is a very simple wrapper around ds.rm to allow you to remove more than 
one object at a time.
}
