% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-discrepancy.R
\name{dh.classDescrepancy}
\alias{dh.classDescrepancy}
\title{Class descrepancy}
\usage{
dh.classDescrepancy(df, vars = NULL)
}
\arguments{
\item{df}{opal dataframe}

\item{vars}{vector of variable names in dataframe (optional). If vars is not 
provided all variables will be included.}
}
\value{
a tibble with columns for (i) variable, (ii) discrepancy (y/n) and
       columns for each cohort indicating the class of the variable
}
\description{
Either due to using ds.dataFrameFill, or because of mistakes in uploading 
data, some cohorts may have variables uploaded as a different class to other
cohorts. This can create problems, e.g. when using ds.summary. This function
produces a table comparing the class of multiple variables.
}
\author{
Tim Cadman
}
