% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-cols.R
\name{dh.columnCast}
\alias{dh.columnCast}
\title{Casts column classes of a table}
\usage{
dh.columnCast(
  df = NULL,
  target_vars = NULL,
  target_class = NULL,
  conns = NULL,
  checks = TRUE
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{target_vars}{Character vector specifying which columns are to be casted to new classes.}

\item{target_class}{Character vector specifying the objective classes of the selected columns.}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the
function. Default is TRUE.}
}
\value{
Tibble with a summary of the successful and failed casts
}
\description{
This function allows you to specify the desired classes for a some variables
of a dataframe. The resulting dataframe will have the same column order as
the input one. If multiple columns are to be casted but only one objective class
is provided, it will be reused for all the selected columns.
}
