% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-rm.R
\name{dh.countRM}
\alias{dh.countRM}
\title{Creates a variable summarising how many observations each subject has in a
dataframe}
\usage{
dh.countRM(df, idvar)
}
\arguments{
\item{df}{opal dataframe}

\item{idvar}{ID variable in that dataset that you want to count}
}
\value{
The original dataframe with an additional variable "n_repeated_obs"
}
\description{
When conducting multilevel modelling we want to subset the data to only 
include subjects with at least two observations. This function creates a
variable summarising this. Note that this is calculated based on the entire
dataframe, so the number provided may be greater than for the specific 
variable you want to model. It is recommended therefore that you first 
subset your dataframe to contain only the repeated measures variable that
you plan to model.
}
\author{
Tim Cadman
}
