% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-cases.R
\name{dh.defineCases}
\alias{dh.defineCases}
\title{Indicates whether each subject has any or all of the variables contained
within a set}
\usage{
dh.defineCases(
  df = NULL,
  vars = NULL,
  type = NULL,
  conns = NULL,
  new_obj = NULL,
  checks = FALSE,
  newobj = NULL
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{vars}{A character vector of columns within `df` to form definition set.}

\item{type}{whther to define cases based on any or all provided variables}

\item{conns}{DataSHIELD connections object.}

\item{new_obj}{optional name for outputted object. Defaults to "dc_data_avail"}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}

\item{newobj}{Retired argument name. Please use `new_obj' instead.}
}
\value{
None. A new variable is created within the opal environment. If the option
        "any" is selected for argument "type", the new variable is called "dc_any_data".
        If the option "all" is selected, the new variable is called "dc_all_data"
}
\description{
In an analysis you may want to subset your dataset to contain only subjects 
meeting a specific criteria, e.g. data on at least one outcome or data on all 
exposures. This function automates this operation by describing whether a 
subject has non-missing values for any or all of a set of given variables.
}
\details{
This function replaces the deprecated dh.subjHasData.
}
