% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-vars-exist.R
\name{dh.doVarsExist}
\alias{dh.doVarsExist}
\title{Check whether variables exist in dataframe}
\usage{
dh.doVarsExist(conns = opals, df, vars)
}
\arguments{
\item{conns}{connections object to DataSHIELD backends}

\item{df}{datashield dataframe}

\item{vars}{vector of variable names expected to be contained in dataframe}
}
\value{
None. Stops function if var(s) don't exist in one of more cohorts.
}
\description{
Probably mostly for internal use, this function can be used at the start
of functions where users provide both vars and a dataframe to check that
the vars exist in the dataframe provided
}
