% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-cols.R
\name{dh.dropCols}
\alias{dh.dropCols}
\title{Subsets a dataframe using column names.}
\usage{
dh.dropCols(
  df = NULL,
  vars = NULL,
  new_obj = df,
  type = c("remove", "keep"),
  conns = NULL,
  checks = TRUE,
  new_df_name = NULL
)
}
\arguments{
\item{df}{datashield dataframe}

\item{vars}{variables to keep or remove}

\item{new_obj}{name for the new dataframe}

\item{type}{whether to remove or keep specified variables}

\item{conns}{connections object to DataSHIELD backends}

\item{checks}{Boolean. Whether or not to perform checks prior to running function. Default is TRUE.}

\item{new_df_name}{Retired argument name. Please use `new_obj' instead.}
}
\value{
a new dataframe is created containing the specified subset of columns
}
\description{
Often we may have created variables within a dataframe that we no longer
need. This function allows you to specify either the variables you want to
keep or drop and create a new dataframe with only the variables you need.
}
