% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-vars-index.R
\name{dh.findVarsIndex}
\alias{dh.findVarsIndex}
\title{Return return indices of columns in opal dataframe}
\usage{
dh.findVarsIndex(df = NULL, vars = NULL, conns = NULL)
}
\arguments{
\item{df}{datashield dataframe}

\item{vars}{vector of variable names in dataframe}

\item{conns}{connections object for DataSHIELD backends}
}
\value{
list of indices where length of list is number of cohorts provided
}
\description{
Some ds functions ask you to provide the index of variables as arguments.
This is highly susceptable to breaking! This is a simple function that will
return the index of variables for a given cohort. It is usefully used in
combination the "keep.cols" argument of ds.dataFrameSubset.
}
