% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-anon-plot-data.R
\name{dh.getAnonPlotData}
\alias{dh.getAnonPlotData}
\title{Extracts an anonymised version of serverside data which can be used to create 
bespoke plots}
\usage{
dh.getAnonPlotData(
  df = NULL,
  v1 = NULL,
  v2 = NULL,
  conns = NULL,
  checks = TRUE
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{v1}{A character giving the name of a column within `df` for which to 
extract anonymised data.}

\item{v2}{An optional second character giving the name of a column within 
`df`. This is only required if you want scatter plot data.}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}
}
\value{
A list of the length of the number of variables provided containing
        anonymised values for each subject of each cohort provided.
}
\description{
Whilst DataSHIELD has plotting functionality, for publications we may need 
more flexiblity in creating plots. This function calls the serverside 
functions which create an anonymised copy of data, and returns the values in
a local object. You can then use these values (e.g. with ggplot2) to create
plots.
}
