% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rm-stats.R
\name{dh.getRmStats}
\alias{dh.getRmStats}
\title{Produces descriptive statistics based on repeated measures data
which it would be useful to report in papers.}
\usage{
dh.getRmStats(df = NULL, outcome = NULL, age_var = NULL, conns = NULL)
}
\arguments{
\item{df}{datashield dataframe}

\item{outcome}{name of outcome variable in df}

\item{age_var}{name of age variable in df}

\item{conns}{connection object for DataSHIELD backends}
}
\description{
Produces descriptive statistics based on repeated measures data
which it would be useful to report in papers.
}
