% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme-mult-poly.R
\name{dh.lmeMultPoly}
\alias{dh.lmeMultPoly}
\title{Function to perform every combination of MLM fractional polynomials}
\usage{
dh.lmeMultPoly(conns = conns, data, outcome, type = "Normal")
}
\arguments{
\item{conns}{connection objects for DataSHIELD backends}

\item{data}{name of dataFrame}

\item{outcome}{list of outcome variables}

\item{type}{type of LME}
}
\description{
Function to perform every combination of MLM fractional polynomials
}
