% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer-tab.R
\name{dh.lmerTab}
\alias{dh.lmerTab}
\title{Extracts the coefficients and confidence intervals from a ds.lmerSLMA model.}
\usage{
dh.lmerTab(fit, coh, coh_format = "cols")
}
\arguments{
\item{fit}{saved output from ds.lmerSLMA model}

\item{coh}{vector of cohorts to include}

\item{coh_format}{whether cohorts are in columns or rows}
}
\value{
A tibble containing the estimate with lower and upper confidence intervals
}
\description{
Extracts the coefficients and confidence intervals from a ds.lmerSLMA model.
}
