% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-tab.R
\name{dh.lmTab}
\alias{dh.lmTab}
\title{Extracts coefficients and confidence intervals from linear models}
\usage{
dh.lmTab(
  model = NULL,
  type = NULL,
  coh_names = NULL,
  direction = NULL,
  ci_format = NULL,
  family = "gaussian",
  digits = 2,
  exponentiate = FALSE,
  extract_random = FALSE
)
}
\arguments{
\item{model}{Object returned from either ds.glm, ds.glmSLMA or ds.lmerSLMA
functions.}

\item{type}{Character specifying type of object provided in \code{model}. Can be
either "glm_ipd", "glm_slma" or "lmer_slma".}

\item{coh_names}{Character vector of cohorts included in \code{model}. Note this
must be in the same order as the cohorts were included in the model.}

\item{direction}{Character specifying the output format. Can be either "wide"
or "long". See "Value" for more details.}

\item{ci_format}{format for the confidence intervals when direction is "wide".
If "separate", upper and lower confidence intervals are displayed as columns.
If "paste", confidence intervals are returned in the same column as the
coefficient within brackets. Option "paste" is only available if \code{ci_format}
is "wide"}

\item{family}{Specifies the family used in the analysis where type is
"glm_ipd" or "glm_slma". Options are "gaussian" or "binomial", with default
"gaussian".}

\item{digits}{Optionally, the number of decimal places to round output to.
Default is 2.}

\item{exponentiate}{Optionally, specify whether estimates from binomial models should
be exponentiated, ie returned as odds ratios. This argument is ignored if
\code{type} is "gaussian".}

\item{extract_random}{Optionally, specify whether to return random effects in mixed effects models.}
}
\value{
A tibble. When \code{direction} is "wide" & \code{ci_format} is "paste", this
contains five columns:
\itemize{
\item variable
\item est
\item lowci
\item uppci
\item pvalue
}

When \code{direction} is "long", a tibble with three columns is returned:
\itemize{
\item variable
\item coefficient (containing values "est", "lowci", "uppci", "pvalue")
\item value
}

When \code{type} is "lmer_slma", a list of 2 elements where the first element
is as described as above, and the second element is a tibble of random
effects with 5 columns: "cohort", "group", "var1", "var2", "stddev".
"stddev" gives either the

#' Extracts the random effects from an lmer_slma model. Returns tibble
containing, for each cohort, the standard deviation of the random terms
and the standard deviation of the residual error. If two or more random
terms are included, it also returns the correlation between the terms.
}
\description{
To conveniently view model results or make tables, it is useful to extract
the coefficients in a useable format. This function extracts coefficients for
ds.glm, ds.glmSLMA and ds.lmerSLMA objects.
}
\seealso{
Other descriptive functions: 
\code{\link{dh.anyData}()},
\code{\link{dh.classDiscrepancy}()},
\code{\link{dh.createTableOne}()},
\code{\link{dh.defineCases}()},
\code{\link{dh.getStats}()},
\code{\link{dh.meanByGroup}()}
}
\concept{descriptive functions}
