% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-proxy.R
\name{dh.localProxy}
\alias{dh.localProxy}
\title{Check whether variables exist in dataframe}
\usage{
dh.localProxy(conns = conns, df)
}
\arguments{
\item{conns}{connections object to DataSHIELD backends}

\item{df}{datashield dataframe}
}
\value{
None. Stops function if var(s) don't exist in one of more cohorts.
}
\description{
Probably mostly for internal use, this function can be used at the start
of functions where users provide both vars and a dataframe to check that
the vars exist in the dataframe provided
}
