% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-iqr.R
\name{dh.makeIQR}
\alias{dh.makeIQR}
\title{Transforms variables based on their interquartile range}
\usage{
dh.makeIQR(
  df = NULL,
  vars = NULL,
  type = c("separate", "pooled"),
  conns = NULL,
  new_obj = df,
  checks = TRUE,
  new_df_name = NULL
)
}
\arguments{
\item{df}{A character giving the name of a server-side data frame.}

\item{vars}{A character vector of columns within `df` to transform.}

\item{type}{either "separate" to transform the variable based on the IQR
calculated within cohort, or "pooled" to transform based on the
pooled IQR across all cohorts provided in the 'conns' argument.}

\item{conns}{connections object to DataSHIELD backends}

\item{new_obj}{name for new dataframe with original vars and iqr versions.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}

\item{new_df_name}{Retired argument name. Please use `new_obj' instead.}
}
\value{
the original dataframe with transformed variables added with the
        suffix "_iqr_c" (if cohort range was used) or "iqr_p" if pooled
        range was used.
}
\description{
This function is used to scale variables by the interquartile range
calulcated either within cohort or using the pooled IQR across cohorts.
The formula is: value(subject) / (75th percentile - 25th percentile).
}
