% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-lmer-form.R
\name{dh.makeLmerForm}
\alias{dh.makeLmerForm}
\title{Make formulae for fitting multiple fractional polynomial models}
\usage{
dh.makeLmerForm(
  outcome = NULL,
  id_var = NULL,
  age_vars = NULL,
  random = NULL,
  fixed = NULL,
  age_interactions = NULL
)
}
\arguments{
\item{outcome}{Character specifying the name of outcome variable within data
frame to be specified in \code{df} argument in \code{dh.lmeMultPoly}.}

\item{id_var}{Character specifying the name of the column which uniquely
identifies each subject within data frame to be specified in \code{df}
argument in \code{dh.lmeMultPoly}.}

\item{age_vars}{Character vector specifying names of age polynomials present
in data frame specified in \code{df} argument in \code{dh.lmeMultPoly}.}

\item{random}{Specifies random effects to include in formulae. Use either
"intercept" for random intercept model or "slope" for random slope model.}

\item{fixed}{Optionally, character vector specifying fixed effects to be
included in model. If no argument is provided default is to create formula
only including outcome, age terms and random effects.}

\item{age_interactions}{Optionally, character vector specifying variable
provided to \code{fixed} for which to create interactions with \verb{age vars}.
If no argument is provided default is to create formula with no interaction
terms.}
}
\value{
Tibble containing two columns:
\itemize{
\item polys = Transformations of \code{age_var}
\item formula = Formula to be used as input to ds.lmerSLMA or dh.lmeMultPoly.
}
}
\description{
This function is designed to be used with \code{dh.lmeMultPoly}. It generates
formulae for multiple fractional polynomial models which can be used as
input to the \code{formulae} argument in \code{dh.lmeMultPoly}.
}
\seealso{
Other trajectory functions: 
\code{\link{dh.lmeMultPoly}()},
\code{\link{dh.makeAgePolys}()},
\code{\link{dh.predictLmer}()},
\code{\link{dh.trimPredData}()},
\code{\link{dh.zByGroup}()}
}
\concept{trajectory functions}
