% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-outcome.R
\name{dh.makeOutcome}
\alias{dh.makeOutcome}
\title{This was the original version of dh.makeStrata. It was been improved to be much more efficient and clearly coded.}
\usage{
dh.makeOutcome(
  df = NULL,
  outcome = NULL,
  age_var = NULL,
  bands = NULL,
  mult_action = NULL,
  mult_vals = NULL,
  keep_original = FALSE,
  df_name = NULL,
  conns = NULL,
  id_var = "child_id",
  band_action = NULL
)
}
\description{
This was the original version of dh.makeStrata. It was been improved to be much more efficient and clearly coded.
}
