% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-by-age.R
\name{dh.meanByAge-defunct}
\alias{dh.meanByAge-defunct}
\title{dh.meanByAge}
\usage{
dh.meanByAge(conns, df, outcome, age_var, intervals, checks)
}
\arguments{
\item{conns}{DataSHIELD connections object.}

\item{df}{Character specifying a server-side data frame.}

\item{outcome}{outcome variable in long format}

\item{age_var}{age in years}

\item{intervals}{table defining our age bands}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}
}
\value{
Mean values for each unit of your age variable are returned
}
\description{
This was the original version of dh.meanByGroup. It is now defunct.
}
\seealso{
\code{\link{dsHelper-deprecated}}
}
\keyword{internal}
