% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-by-occasion.R
\name{dh.meanByOccasion}
\alias{dh.meanByOccasion}
\title{Derives one or more outcome variable(s) from repeated measures data}
\usage{
dh.meanByOccasion(df, agevar, outcome, bands, conns, grouping = NULL)
}
\arguments{
\item{df}{serverside dataframe}

\item{agevar}{variable indicating subject age/time}

\item{outcome}{outcome variable}

\item{bands}{vector of even length given the upper and low age bands to 
create the outcomes at.}

\item{conns}{connections object for DataSHIELD backends}

\item{grouping}{an optional variable if you want to stratify the means by 
another variable, e.g. socioeconomic position.}
}
\value{
a tibble containing mean values of the outcome at different age bands
}
\description{
When running multilevel models we often want to visualise the actual data,
for example the average outcome at different time points. This is useful to
check visually how well our model fits the data. This function does that. At 
the moment this is a really slow, convoluted way to do it involving taking 
lots of subsets. This is because tapply.assign doesn't work properly at the
moment.
}
