% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-sep-models.R
\name{dh.metaSepModels}
\alias{dh.metaSepModels}
\title{Function in progress to meta-analyse separate models.}
\usage{
dh.metaSepModels(
  input = "fit",
  ref = NULL,
  exp = NULL,
  method = NULL,
  output = "both"
)
}
\arguments{
\item{input}{Character; either "fit" if input is glm object, or "coefs" if input is table of coefficients returned by dh.lmTab.}

\item{ref}{Tibble, output from dh.multGlm.}

\item{exp}{Logical, whether to exponentiate coefficients after meta-analysis}

\item{method}{Method of meta-analysing coefficients.}

\item{output}{Character; "cohort" to return cohort coefficients only, "meta"
to return meta-analysed coefficients only, "both" to return both.}
}
\value{
A tibble
}
\description{
Function in progress to meta-analyse separate models.
}
\concept{model building}
