% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not-completely-missing.R
\name{dh.notCompletelyMissing}
\alias{dh.notCompletelyMissing}
\title{Not completely missing variables}
\usage{
dh.notCompletelyMissing(conns = opals, df)
}
\arguments{
\item{conns}{connections object for DataSHIELD backends}

\item{df}{sever side dataframe}
}
\value{
a dataframe with columns for each variable and rows for each cohort indicating if the variable is not completely missing
}
\description{
When performing a Study Level Meta Analysis, the intention should be that a consistent set of confounding variables for each cohort.
However it is not strictly necessary to have all confounding variables available in all cohorts. Assume that some cohorts
have completely empty confounding variables. In this scenario it is useful
to be able to specify a model to fit on all cohorts and use a look up table to understand if there are any cohorts that lack a
particular confounder. The purpose of this function is to generate this look up table automatically.
}
