% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{dh.pool}
\alias{dh.pool}
\title{Perform Rubin's pooling on a list of imputed generalized linear models.}
\usage{
dh.pool(
  imputed_glm = NULL,
  type = NULL,
  coh_names = NULL,
  family = NULL,
  exponentiate = FALSE
)
}
\arguments{
\item{imputed_glm}{A list of imputed generalized linear models.}

\item{type}{Character specifying type of object provided in `model`. Can be
either "glm_ipd", "glm_slma" or "lmer_slma".}

\item{coh_names}{Character vector of cohorts included in `imputed_glm`. Note this
must be in the same order as the cohorts were included in the model.}

\item{family}{Specifies the family used in the analysis where type is
"glm_ipd" or "glm_slma". Options are "gaussian" or "binomial", with default
"gaussian".}

\item{exponentiate}{Optionally, specify whether estimates from binomial models should
be exponentiated, ie returned as odds ratios. This argument is ignored if
`type` is "gaussian".}
}
\value{
A tibble containing Rubin's pooled estimates and confidence intervals.
}
\description{
Perform Rubin's pooling on a list of imputed generalized linear models.
}
\details{
This function performs Rubin's pooling on a list of imputed generalized linear models. 
It extracts coefficients using specified parameters, tidies the coefficients, and then performs 
pooling.
}
