% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartile-split.R
\name{dh.quartileSplit}
\alias{dh.quartileSplit}
\title{Splits a continuous variables into four quartiles}
\usage{
dh.quartileSplit(
  df = NULL,
  var = NULL,
  new_obj = NULL,
  band_action = NULL,
  type = NULL,
  var_suffix = "_q_",
  conns = NULL
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{var}{Character specifying continuous variable to transform into
quartiles.}

\item{new_obj}{Character specifying name for created serverside object.}

\item{band_action}{Character specifying how the quartiles are separated:
\itemize{
\item "g_l" = greater than the lowest band and less than the highest band
\item "ge_le" = greater or equal to the lowest band and less than or equal to the
highest band
\item "g_le" = greater than the lowest band and less than or equal to the highest
band
\item "ge_l" = greater than or equal to the lowest band and less than the highest
band
}}

\item{type}{Character specifying whether to derive quartiles from combined
data or within each cohort. Use "combine" to use combined quartiles, and
"split" to use cohort-specific quartiles.}

\item{var_suffix}{Character specifying the suffix to give the created variable.
Default is "\emph{q}"}

\item{conns}{DataSHIELD connections object.}
}
\value{
Servside dataframe in containing a maximum of four additional
variables representing the quantiles of the original variable. If a cohort
has insufficient observations within that quartile (less than the filter
threshold) the variable will not be created an a warning will be returned.
}
\description{
For some sensitivity analyses you may want to examine potential non-linear
associations between an exposure and outcome. This function splits a
continuous variable into four continuous variables representing each
quartile of the original variable. Participants with no measurement within
a quartile will be assigned NA.
}
\seealso{
Other data manipulation functions: 
\code{\link{dh.dropCols}()},
\code{\link{dh.makeAgePolys}()},
\code{\link{dh.makeIQR}()},
\code{\link{dh.makeStrata}()},
\code{\link{dh.renameVars}()},
\code{\link{dh.tidyEnv}()}
}
\concept{data manipulation functions}
