% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-vars.R
\name{dh.renameVars}
\alias{dh.renameVars}
\title{Rename one or more columns within a serverside data frame}
\usage{
dh.renameVars(
  df = NULL,
  current_names = NULL,
  new_names,
  conns = NULL,
  checks = TRUE
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{current_names}{Character vector of columns within \code{df} to rename.}

\item{new_names}{Character vector giving the new names for the columns
specified in \code{current_names}.}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the
function. Default is TRUE.}
}
\value{
Data frame specified in \code{df} is returned server-side with variables
renamed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was an analogue of \code{dplyr::rename()}. It was deprecated because you can now use
\code{dsTidyverseClient::ds.rename()} which is much quicker and more flexible.
}
\keyword{internal}
