% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablised-weights.R
\name{dh.stablisedWeights}
\alias{dh.stablisedWeights}
\title{Generate Stabilized Weights Using ds.glmSLMA output and outcome proportions.}
\usage{
dh.stablisedWeights(
  glm_object = NULL,
  outcome_df = NULL,
  outcome_var = NULL,
  new_obj = NULL,
  conns = NULL
)
}
\arguments{
\item{glm_object}{A fitted ds.glmSLMA object.}

\item{outcome_df}{A serverside data frame containing the outcome variable.}

\item{outcome_var}{Character vector of outcome variable within `outcome_df`.}

\item{new_obj}{A character string for the name of the new object to store stabilized weights.}

\item{conns}{DataSHIELD connections object.}
}
\value{
Stabilized weights assigned as serverside object `new_obj`
}
\description{
This calculates stablised inverse probability weights. Currently only works with a binary
outcome variable.
}
\details{
The function performs the following steps:
  1. Generate predicted values from the GLM model.
  2. Calculate proportions of a binary outcome using the outcome data.
  3. Format the proportions data for analysis.
  4. Create stabilized weights using the formatted proportions and GLM predictions.
}
\note{
Currently only works with binary outcome variable.
}
