% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subj-has-data.R
\name{dh.subjHasData}
\alias{dh.subjHasData}
\title{This was an old version of dh.defineCases which is now defunct}
\usage{
dh.subjHasData()
}
\description{
This was an old version of dh.defineCases which is now defunct
}
