% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-env.R
\name{dh.tidyEnv}
\alias{dh.tidyEnv}
\title{Remove multiple objects from the serverside environment}
\usage{
dh.tidyEnv(obj = NULL, type = NULL, conns = NULL)
}
\arguments{
\item{obj}{objects that you want to either keep or remove}

\item{type}{either "remove" to remove the listed objects of "keep" to keep
the listed objects and remove everything else.}

\item{conns}{DataSHIELD connections object.}
}
\value{
None. Objects removed from ds environment
}
\description{
This is a wrapper around ds.rm to allow you to remove multiple objects 
in one call.
}
