% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-pred-data.R
\name{dh.trimPredData}
\alias{dh.trimPredData}
\title{Trims predicted values based on min and max values provided}
\usage{
dh.trimPredData(
  pred = NULL,
  coh_names = NULL,
  age = "age",
  min = NULL,
  max = NULL
)
}
\arguments{
\item{pred}{Tibble of predicted values returned by dh.predictLmer.}

\item{coh_names}{Character Vector of cohort names corresponding to cohorts
included in `pred`.}

\item{age}{Character specifying name of the age variable in `pred`.}

\item{min}{Vector of minimum values of age with length equal to number of
cohorts in `pred`.}

\item{max}{Vector of maximum values of age with length equal to number of
cohorts in `pred`.}
}
\value{
`pred` is returned as a tibble with predicted values outside of
`min` and `max` removed.
}
\description{
This is to be used in conjuncture with `ds.predictLmer`. The new data frame
used for prediction may contain age or time values not present in the
original data. This allows you to remove all predicted values beyond a
minimum and maximum age. This ensures that plots only show predict values
over a range where there was actually data.
}
\seealso{
Other trajectory functions: 
\code{\link{dh.lmeMultPoly}()},
\code{\link{dh.makeAgePolys}()},
\code{\link{dh.makeLmerForm}()},
\code{\link{dh.predictLmer}()},
\code{\link{dh.zByGroup}()}
}
\concept{trajectory functions}
