% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-score-by-group.R
\name{dh.zByGroup}
\alias{dh.zByGroup}
\title{Creates z-scores within specified bands}
\usage{
dh.zByGroup(df, out_var, age_var, low_band, upp_band, conns, new_obj)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{out_var}{Variable to make z-scores for}

\item{age_var}{Age variable}

\item{low_band}{Lower band for z score}

\item{upp_band}{Upper band for z score}

\item{conns}{DataSHIELD connections object.}

\item{new_obj}{Character specifying name for created serverside object.}
}
\value{
Z score serverside within specified bands
}
\description{
Especially with mental health outcomes, we often want to transform raw scores
into z-scores, but within certain age bands (e.g. measurement occasion or
per year). This function does this.
}
\seealso{
Other trajectory functions: 
\code{\link{dh.lmeMultPoly}()},
\code{\link{dh.makeAgePolys}()},
\code{\link{dh.makeLmerForm}()},
\code{\link{dh.predictLmer}()},
\code{\link{dh.trimPredData}()}

Other data manipulation functions: 
\code{\link{dh.dropCols}()},
\code{\link{dh.makeAgePolys}()},
\code{\link{dh.makeIQR}()},
\code{\link{dh.makeStrata}()},
\code{\link{dh.quartileSplit}()},
\code{\link{dh.renameVars}()},
\code{\link{dh.tidyEnv}()}
}
\concept{data manipulation functions}
\concept{trajectory functions}
