% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-table-one.R
\name{.checkLabCols}
\alias{.checkLabCols}
\title{Checks that the columns provided in labs contain all and only those provided
in `required_cols`}
\usage{
.checkLabCols(labs, n_col, required_cols)
}
\arguments{
\item{labs}{Object provided either to var_labs, cat_labs or coh_labs}

\item{n_col}{Integer specifying the required number of columns}

\item{required_cols}{Character vector specifying the required column names}
}
\value{
Returns an error if number and names of columns does not match that
specified

noRd
}
\description{
Checks that the columns provided in labs contain all and only those provided
in `required_cols`
}
