% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Initialization functions to do init actions}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{pkgname}{package name}

\item{name}{of the library}
}
\description{
Initialization functions to do init actions
}
\keyword{internal}
