% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsHelper-defunct.R, R/mean-by-age.R
\name{dsHelper-defunct}
\alias{dsHelper-defunct}
\alias{dh.meanByAge}
\title{Defunct functions in package \pkg{dsHelper}.}
\usage{
dh.meanByAge(...)
}
\description{
The functions listed below are defunct and no longer available.
When possible, alternative functions with similar functionality are
mentioned. Help pages for deprecated functions are available at
\code{help("<function>-defunct")}.
}
\section{\code{dh.meanByAge}}{

For \code{dh.meanByAge}, use \code{\link[=dh.meanByGroup]{dh.meanByGroup()}}.
}

\keyword{internal}
