% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsHelper-package.R
\docType{package}
\name{dsHelper-package}
\alias{dsHelper}
\alias{dsHelper-package}
\title{dsHelper: Helper Functions for Use with 'DataSHIELD'}
\description{
Often we need to automate things with 'DataSHIELD'. These functions help to do that.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lifecycle-project/ds-helper/}
  \item \url{https: //lifecycle-project.github.io/ds-helper/}
  \item Report bugs at \url{https://github.com/lifecycle-project/ds-helper/issues/}
}

}
\author{
\strong{Maintainer}: Tim Cadman \email{t.cadman@bristol.ac.uk} (\href{https://orcid.org/0000-0002-7682-5645}{ORCID})

Other contributors:
\itemize{
  \item Sido Haakma \email{s.haakma@rug.nl} (\href{https://orcid.org/0000-0003-1278-9144}{ORCID}) [contributor]
  \item Tom Bishop \email{tom.bishop@mrc-epid.cam.ac.uk} [contributor]
}

}
\keyword{internal}
