% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-excluded-df.R
\name{dt.makeExcludedDf}
\alias{dt.makeExcludedDf}
\title{Given df-A & df-B, creates a new df which is the rows in A but not in B}
\usage{
dt.makeExcludedDf(
  original_df,
  final_df,
  id_var = "child_id",
  new_obj,
  type = "wide",
  conns = NULL
)
}
\arguments{
\item{original_df}{Dataframe containing the full sample}

\item{final_df}{Dataframe containing the included sample}

\item{new_obj}{Character specifying name for created serverside object.}

\item{type}{Specifies type of dataframe in `original_df`. Either 'long' or
'wide'. NOTE NOT CURRENTLY FUNCTIONAL - ONLY WORKS FOR WIDE.}

\item{conns}{DataSHIELD connections object.}
}
\value{
Creates a serverside dataframe containing the rows from `original_df`
that are not contained in `final_df`
}
\description{
When writing a paper often we need to exclude various participants for 
various reasons. Then we will need a df with all of these excluded 
participants. This is one way to do it.
}
