% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any-data.R
\name{dh.anyData}
\alias{dh.anyData}
\title{Describe whether variables are completely missing for each cohort}
\usage{
dh.anyData(df = NULL, vars = NULL, conns = NULL, checks = TRUE)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{vars}{Character vector of columns within `df`.}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}
}
\value{
Client-side tibble containing columns for each variable and rows
for each cohort indicating if the variable is not completely missing.
}
\description{
When performing a Study Level Meta Analysis of coefficients, it
is usually necessary that all cohorts have some data on all variables in the
model. Manually identifying which cohorts have some data on required
variables is time-consuming and at risk of error. This function automatically
generates a look-up table with this information.
}
\seealso{
Other descriptive functions: 
\code{\link{dh.classDiscrepancy}()},
\code{\link{dh.defineCases}()},
\code{\link{dh.getStats}()},
\code{\link{dh.lmTab}()},
\code{\link{dh.meanByGroup}()}
}
\concept{descriptive functions}
