% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-models.R
\name{dh.buildModels}
\alias{dh.buildModels}
\title{This function is part of a pipeline to build regression models in DataSHIELD}
\usage{
dh.buildModels(avail_exp = NULL, avail_cov = NULL, outcome = NULL)
}
\arguments{
\item{avail_exp}{Tibble, output from dh.anyData. Should contain all 
exposures you want to use. One model will be created for each exposure.}

\item{avail_cov}{Tibble, output from dh.anyData. Should contain all 
covariates to include in your models.}

\item{outcome}{Character giving the name of your outcome variable.}
}
\value{
A tibble with five columns: exposure, cohort, covariates, outcome,
formula.

Currently this just builds formula for ds.lmerSLMA, but I should probably 
try and combine it with the lmer loop function.
}
\description{
When working in DataSHIELD you often end up finding hundreds of slightly
different models. For example, you might want to include differen covariates
for different cohorts, or perform sensitivity analyses with slight 
differences to the main model. This function helps with that. I will describe
the flow in more detail in a separate vignette, but here it is briefly:
}
\details{
dh.anyData: identify available exposures and outcomes in your data
dh.buildModels: based on this, build models where data is available
dh.multGlm: fit multiple models and store the output
dh.metaSepModels: meta-analyse the coefficients and return key output in a 
neat table
}
\seealso{
Other model building: 
\code{\link{dh.metaSepModels}()}
}
\concept{model building}
