% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-discrepancy.R
\name{dh.classDiscrepancy}
\alias{dh.classDiscrepancy}
\title{Describes the class of one or more variables across cohorts and indicates 
differences}
\usage{
dh.classDiscrepancy(df = NULL, vars = NULL, conns = NULL, checks = TRUE)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{vars}{Optionally, a character vector specifying columns within `df` to 
describe. If NULL all variables will be included.}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}
}
\value{
A tibble with columns for each variable and rows for each cohort 
describing the class of the variable, with an additional column 'discrepancy'
summarising whether there are differences between cohorts.
}
\description{
Either through using ds.dataFrameFill, or due to mistakes in uploading
data, the same variable may have a different class across cohorts. This can 
create problems as many DataSHIELD functions require the input variable to 
have the same class in all studies. This function produces a tibble showing 
the class of each variable.
}
\seealso{
Other descriptive functions: 
\code{\link{dh.anyData}()},
\code{\link{dh.defineCases}()},
\code{\link{dh.getStats}()}
}
\concept{descriptive functions}
