% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-cols.R
\name{dh.columnCast}
\alias{dh.columnCast}
\title{Casts column classes of a table}
\usage{
dh.columnCast(
  df = NULL,
  target_vars = NULL,
  target_class = NULL,
  conns = NULL,
  checks = TRUE
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{target_vars}{Character vector specifying which columns are to be casted to new classes.}

\item{target_class}{Character vector specifying the objective classes of the selected columns.}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the
function. Default is TRUE.}
}
\value{
Tibble with a summary of the successful and failed casts
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function allowed you to change the class of variables within a data frame. It was depricated
because you can now use \code{dsTidyverseClient::ds.mutate()} which is much quicker and has greater
flexibility.
}
\keyword{internal}
