% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-cases.R
\name{dh.defineCases}
\alias{dh.defineCases}
\title{Indicates whether each subject has any or all of the variables contained
within a set}
\usage{
dh.defineCases(
  df = NULL,
  vars = NULL,
  type = NULL,
  conns = NULL,
  new_obj = NULL,
  checks = FALSE,
  newobj = NULL
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{vars}{Character vector of columns within `df` to form definition set.}

\item{type}{Character specifying how to define cases. Use "any" to describe
subjects with data on any of `vars`, and "all" to describe subjects with data 
on all of `vars.}

\item{conns}{DataSHIELD connections object.}

\item{new_obj}{Character specifying name for created serverside object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}

\item{newobj}{Retired argument name. Please use `new_obj' instead.}
}
\value{
Server-side vector defining whether subject meets criteria defined by
`vars` and `type`. 1 indicates that criteria were met, 0 indicates that 
criteria weren't met.
}
\description{
In an analysis you may want to subset your dataset to contain only subjects 
meeting a specific criteria, e.g. data on at least one outcome or data on all 
exposures. This function automates this operation by describing whether a 
subject has non-missing values for any or all of a set of given variables.
}
\details{
This function replaces the deprecated dh.subjHasData.
}
\seealso{
Other descriptive functions: 
\code{\link{dh.anyData}()},
\code{\link{dh.classDiscrepancy}()},
\code{\link{dh.getStats}()},
\code{\link{dh.lmTab}()},
\code{\link{dh.meanByGroup}()}
}
\concept{descriptive functions}
