% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-cases.R
\name{dh.defineCases}
\alias{dh.defineCases}
\title{Indicates whether each subject has any or all of the variables contained
within a set}
\usage{
dh.defineCases(
  df = NULL,
  vars = NULL,
  type = NULL,
  conns = NULL,
  newobj = "dc_data_avail"
)
}
\arguments{
\item{df}{datashield dataframe}

\item{vars}{vector of variable names in dataframe}

\item{type}{whther to define cases based on any or all provided variables}

\item{conns}{connection object for DataSHIELD backends}

\item{newobj}{optional name for outputted object. Defaults to "dc_data_avail"}
}
\value{
None. A new variable is created within the opal environment. If the option
        "any" is selected for argument "type", the new variable is called "dc_any_data".
        If the option "all" is selected, the new variable is called "dc_all_data"
}
\description{
At some point in the analysis you will want to subset your dataset to contain
only subjects meeting some criteria, e.g. data on at least one outcome or
data on all exposures. This function speeds things up by indicating whether a subject
has non-missing values for any or all of a set of given variables.
}
\details{
Note this function replaces the deprecated dh.subjHasData.
}
