% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-cols.R
\name{dh.dropCols}
\alias{dh.dropCols}
\title{Subsets a dataframe using column names.}
\usage{
dh.dropCols(
  df = NULL,
  vars = NULL,
  new_df_name = df,
  type = c("remove", "keep"),
  conns = NULL
)
}
\arguments{
\item{df}{datashield dataframe}

\item{vars}{variables to keep or remove}

\item{new_df_name}{name for the new dataframe}

\item{type}{whether to remove or keep specified variables}

\item{conns}{connections object to DataSHIELD backends}

\item{comp_var}{name of a variable with df which is complete for all subjects
(e.g. an id variable)}

\item{remove_temp}{remove temporary objects from the DataSHIELD backends}
}
\value{
a new dataframe is created containing the specified subset of columns
}
\description{
Often we may have created variables within a dataframe that we no longer
need. This function allows you to specify either the variables you want to
keep or drop and create a new dataframe with only the variables you need.
}
