% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-cols.R
\name{dh.dropCols}
\alias{dh.dropCols}
\title{Removes columns from a serverside data frame}
\usage{
dh.dropCols(
  df = NULL,
  vars = NULL,
  new_obj = df,
  type = NULL,
  conns = NULL,
  checks = TRUE,
  new_df_name = NULL
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{vars}{Character vector specifying columns within `df` to be removed or
kept.}

\item{new_obj}{Optionally, character specifying name for new server-side
data frame. Default is to return original data frame with columns removed.}

\item{type}{Character specifying how to treat `vars`. If "remove" these
variables are removed from the data frame, if "keep" these variables are
kept in the data frame and all others are removed.}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}

\item{new_df_name}{Retired argument name. Please use `new_obj' instead.}
}
\value{
Server-side data frame the specified subset of columns.
}
\description{
Often we want to remove variables from a dataframe. This function
allows you to specify the variables you either want to remove or keep and
and creates a new dataframe with only the required variables.
}
\seealso{
Other data manipulation functions: 
\code{\link{dh.makeAgePolys}()},
\code{\link{dh.makeIQR}()},
\code{\link{dh.makeStrata}()},
\code{\link{dh.quantileSplit}()},
\code{\link{dh.renameVars}()},
\code{\link{dh.tidyEnv}()}
}
\concept{data manipulation functions}
