% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-cols.R
\name{dh.dropCols}
\alias{dh.dropCols}
\title{Removes columns from a serverside data frame}
\usage{
dh.dropCols(
  df = NULL,
  vars = NULL,
  new_obj = df,
  type = c("remove", "keep"),
  conns = NULL,
  checks = TRUE,
  new_df_name = NULL
)
}
\arguments{
\item{df}{A character giving the name of a server-side data frame.}

\item{vars}{A character vector of columns within `df` to be removed or kept.}

\item{new_obj}{name for the new dataframe}

\item{type}{whether to remove or keep specified variables}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}

\item{new_df_name}{Retired argument name. Please use `new_obj' instead.}
}
\value{
a new dataframe is created containing the specified subset of columns
}
\description{
Often we was to remove variables from a dataframe that we no longer need. 
This function allows you to specify the variables you either want to
keep or drop, and creates a new dataframe with only the required variables.
}
