% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-vars-index.R
\name{dh.findVarsIndex}
\alias{dh.findVarsIndex}
\title{Return return indices of column names in server-side dataframe}
\usage{
dh.findVarsIndex(df = NULL, vars = NULL, conns = NULL, checks = TRUE)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{vars}{Character vector of columns within `df` for which to return the
indices.}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}
}
\value{
Client-side list of indices corresponding to `vars`, where the length
of the list corresponds to the number of cohorts included in `conns`.
}
\description{
Some DataSHIELD functions require column indices as parameters. This is
hqighly susceptable to breaking as changes in code will change the order of
variables. This function allows you to specify the names of columns and
returns their indices.
}
