% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-anon-plot-data.R
\name{dh.getAnonPlotData}
\alias{dh.getAnonPlotData}
\title{Extracts an anonymised version of the dataset which can be used
to create bespoke plots (e.g. using ggplot). This calls the ds.scatterPlot
function but stores the data produced this function. Note you need to make
sure your plot window is large enough to produce the plot.}
\usage{
dh.getAnonPlotData(df = NULL, vars = NULL, conns = NULL)
}
\arguments{
\item{df}{opal dataframe}

\item{vars}{vector of variable names in dataframe}

\item{conns}{connection object for DataSHIELD backends}
}
\value{
A list of the length of the number of variables provided containing
        anonymised values for each subject of each cohort provided.
}
\description{
Extracts an anonymised version of the dataset which can be used
to create bespoke plots (e.g. using ggplot). This calls the ds.scatterPlot
function but stores the data produced this function. Note you need to make
sure your plot window is large enough to produce the plot.
}
