% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rm-stats.R
\name{dh.getRmStats}
\alias{dh.getRmStats}
\title{Produces descriptive statistics based on repeated measures data
which it would be useful to report in papers.}
\usage{
dh.getRmStats(
  df = NULL,
  outcome = NULL,
  id_var = NULL,
  age_var = NULL,
  conns = NULL
)
}
\arguments{
\item{df}{datashield dataframe}

\item{outcome}{name of outcome variable in df}

\item{id_var}{name of id variable in df}

\item{age_var}{name of age variable in df}

\item{conns}{connection object for DataSHIELD backends}
}
\value{
a tibble containing the following columns:

min_age: 5th percentile of age
max_age: 95th percentile of age
n_obs: total number of observations in data
n_participants: total number of unique participants
n_meas_5: 5th percentile of measurements per individual
n_meas_med: median number of measurements per individual
n_meas_95: 95th percentile of measurements per individual
}
\description{
Produces descriptive statistics based on repeated measures data
which it would be useful to report in papers.
}
