% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stats.R
\name{dh.getStats}
\alias{dh.getStats}
\title{Produces a range of descriptive statistics in a useful format}
\usage{
dh.getStats(df = NULL, vars = NULL, digits = 2, conns = NULL, checks = TRUE)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{vars}{Character vector of columns within \code{df} to summarise.}

\item{digits}{Optionally, the number of decimal places to round output to. 
Default is 2.}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}
}
\value{
A client-side list with two elements: "categorical" and "continuous".
Each element contains a tibble with descriptive statistics as follows.

Categorical: \cr
\itemize{
\item "variable" = Variable name.
\item cohort = Cohort name, where "combined" refers to pooled statistics.
\item category = Level of variable, including 'missing' as a category.
\item value = Number of observations within category.
\item cohort_n = Total number of observations per cohort within \code{df}.
\item valid_n = Number of valid observations for variable (sum of ns for all
categories excluding missing).
\item missing_n = Number of missing observations.
\item perc_valid = Numnber of observations within a category as percentage of
valid_n.
\item perc_total = Number of observations within a category as percentage of
cohort_n.
}

Continuous: \cr
\itemize{
\item variable = As above.
\item cohort = As above.
\item mean = Mean. The pooled value calculated by fixed-effect meta-analysis.
\item std.dev = Standard deviation. The pooled value is also calculate by fixed-
\item effect meta-analysis.
\item perc_5, perc_10, perc_25, perc_50, perc_75, perc_90, perc_95 = 5th to 95th
\item percentile values.
\item valid_n = As above.
\item cohort_n = As above.
\item missing_n = As above.
\item missing_perc = Percentage of observations missing.
}
}
\description{
Whilst dsBaseClient provides the functions 'ds.table' and 'ds.summary' to
calculate descriptive statistics, their output is not in a very useable
format. This function extracts key descriptive statistics and returns them in
tibble.
}
\details{
This function also overcomes an issue with ds.summary, where it throws an
error if the variable is missing in one or more study. By contrast,
dh.getStats will return the variable for that cohort with all NAs. See
'value' for details of returned statistics.
}
\seealso{
Other descriptive functions: 
\code{\link{dh.anyData}()},
\code{\link{dh.classDiscrepancy}()},
\code{\link{dh.defineCases}()},
\code{\link{dh.lmTab}()},
\code{\link{dh.meanByGroup}()}
}
\concept{descriptive functions}
