% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme-mult-poly.R
\name{dh.lmeMultPoly}
\alias{dh.lmeMultPoly}
\title{Function to perform every combination of MLM fractional polynomials}
\usage{
dh.lmeMultPoly(
  df = NULL,
  formulae = NULL,
  poly_names = NULL,
  conns = NULL,
  checks = TRUE
)
}
\arguments{
\item{df}{name of dataFrame}

\item{formulae}{a vector of model formulae to fit}

\item{poly_names}{a vector of names for your models}

\item{conns}{connection objects for DataSHIELD backends}

\item{checks}{Boolean. Whether or not to perform checks prior to running function. Default is TRUE.}
}
\description{
Function to perform every combination of MLM fractional polynomials
}
\author{
Tim Cadman
}
