% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme-mult-poly.R
\name{dh.lmeMultPoly}
\alias{dh.lmeMultPoly}
\title{Function to perform every combination of MLM fractional polynomials}
\usage{
dh.lmeMultPoly(conns = conns, df, formulae)
}
\arguments{
\item{conns}{connection objects for DataSHIELD backends}

\item{df}{name of dataFrame}

\item{formulae}{tibble containing formulae with column labelled 'formulae'}
}
\description{
Function to perform every combination of MLM fractional polynomials
}
