% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme-mult-poly.R
\name{dh.lmeMultPoly}
\alias{dh.lmeMultPoly}
\title{Function to perform every combination of MLM fractional polynomials}
\usage{
dh.lmeMultPoly(df = NULL, formulae = NULL, conns = NULL)
}
\arguments{
\item{df}{name of dataFrame}

\item{formulae}{tibble containing formulae with column labelled 'formulae'}

\item{conns}{connection objects for DataSHIELD backends}
}
\description{
Function to perform every combination of MLM fractional polynomials
}
\author{
Tim Cadman
}
