% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme-mult-poly.R
\name{dh.lmeMultPoly}
\alias{dh.lmeMultPoly}
\title{Fit multiple mixed effects models containing different combination of 
fractional polynomials}
\usage{
dh.lmeMultPoly(
  df = NULL,
  formulae = NULL,
  poly_names = NULL,
  conns = NULL,
  checks = TRUE
)
}
\arguments{
\item{df}{A character giving the name of a server-side data frame.}

\item{formulae}{a vector of model formulae to fit}

\item{poly_names}{a vector of names for your models}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}
}
\description{
This function enables you to fit multiple models with different combinations 
of polynomial terms and compares the fit.
}
\author{
Tim Cadman
}
