% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-tab.R, R/lm-tab.R
\name{dh.lmTab}
\alias{dh.lmTab}
\title{Extracts the coefficients and confidence intervals from a ds.glm or
ds.glmSLMA model.}
\usage{
dh.lmTab(model, type, coh_names, direction = "long", ci_format = "paste")

dh.lmTab(model, type, coh_names, direction = "long", ci_format = "paste")
}
\arguments{
\item{model}{saved output from either ds.glm,  ds.glmSLMA or ds.lmerSLMA}

\item{type}{either "ipd", "slma" or "lmer"}

\item{coh_names}{a vector of cohort names. Note this needs to be in the same
order as the cohorts provided to the model.}

\item{direction}{either "long" or "wide"}

\item{ci_format}{format for the confidence intervals when direction == "wide".
"separate" outputs separate columns with upper and lower CIs. 
"paste" adds these in brackets to the coefficient.'}
}
\value{
A tibble containing the estimate with lower and upper confidence intervals

A tibble containing the estimate with lower and upper confidence intervals
}
\description{
Extracts the coefficients and confidence intervals from a ds.glm or
ds.glmSLMA model.

Extracts the coefficients and confidence intervals from a ds.glm or
ds.glmSLMA model.
}
