% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-age-polys.R
\name{dh.makeAgePolys}
\alias{dh.makeAgePolys}
\title{Produces multiple transformations of age term for fractional polynomial
analysis.}
\usage{
dh.makeAgePolys(
  df = NULL,
  age_vars = NULL,
  conns = NULL,
  poly_form = c("^-2", "^-1", "^-0.5", "log", "^0.5", "^2", "^3"),
  poly_names = c("m_2", "m_1", "m_0_5", "log", "0_5", "2", "3"),
  checks = TRUE,
  agevars = NULL
)
}
\arguments{
\item{df}{opal dataframe}

\item{age_vars}{the age variable to transform}

\item{conns}{connections object for DataSHIELD backends}

\item{poly_form}{a vector of powers by which to transform the age variable}

\item{poly_names}{a vector of names for the created variables, the same length
and order as poly_form}

\item{checks}{Boolean. Whether or not to perform checks prior to running function. Default is TRUE.}

\item{agevars}{Retired argument name. Please use `new_obj' instead.}
}
\value{
transformations of age created in df
}
\description{
When we do trajectory analyses using fractional polynomials we often want
to try models with different combinations of age polynomials. This function
creates the following age transformations in datashield: age^-2, age^-1,
age^-0.5, log(age), age^0.5, age^2, age^3.
}
