% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-age-polys.R
\name{dh.makeAgePolys}
\alias{dh.makeAgePolys}
\title{Produces multiple transformations of the age term for fractional polynomial
analyses}
\usage{
dh.makeAgePolys(
  df = NULL,
  age_var = NULL,
  conns = NULL,
  poly_form = c("^-2", "^-1", "^-0.5", "log", "^0.5", "^2", "^3"),
  poly_names = c("_m_2", "_m_1", "_m_0_5", "log", "_0_5", "_2", "_3"),
  checks = TRUE,
  agevars = NULL
)
}
\arguments{
\item{df}{A character giving the name of a server-side data frame.}

\item{age_var}{A character giving the age variable within `df` to transform.}

\item{conns}{DataSHIELD connections object.}

\item{poly_form}{a vector of powers by which to transform the age variable}

\item{poly_names}{a vector of names for the created variables, the same length
and order as poly_form}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}

\item{agevars}{Retired argument name. Please use `new_obj' instead.}
}
\value{
transformations of age created in df
}
\description{
When we do trajectory analyses using fractional polynomials we often want
to try models with different combinations of age polynomials. This function
creates multiple transformations of an age terms and adds the transformed
variables to the original dataframe.
}
