% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-age-polys.R
\name{dh.makeAgePolys}
\alias{dh.makeAgePolys}
\title{Produces multiple transformations of age term for fractional polynomial
analysis.}
\usage{
dh.makeAgePolys(df, agevars, conns = NULL)
}
\arguments{
\item{df}{opal dataframe}

\item{agevars}{the age variable to transform}

\item{conns}{connections object for DataSHIELD backends}
}
\value{
transformations of age created in df
}
\description{
When we do trajectory analyses using fractional polynomials we often want
to try models with different combinations of age polynomials. This function
creates the following age transformations in datashield: age^-2, age^-1,
age^-0.5, log(age), age^0.5, age^2, age^3.
}
