% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-iqr.R
\name{dh.makeIQR}
\alias{dh.makeIQR}
\title{Transforms variables based on their interquartile range}
\usage{
dh.makeIQR(
  df = NULL,
  vars = NULL,
  type = c("separate", "pooled"),
  conns = NULL,
  new_obj = df,
  checks = TRUE,
  new_df_name = NULL
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{vars}{Character vector of columns within `df` to transform.}

\item{type}{Use "separate" to transform the variable based on the 
within-cohort IQR, or "pooled" to use the pooled IQR across all cohorts 
specified in `conns`.}

\item{conns}{DataSHIELD connections object.}

\item{new_obj}{Character specifying name for created serverside object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}

\item{new_df_name}{Retired argument. Please use `new_obj' instead.}
}
\value{
Server-side object specified in `df` with transformed variables added 
as columns. Variables have suffic "_iqr_c" if type is "separate" and suffix
"iqr_p" if pooled is type is "pooled".
}
\description{
This function scales variables by their interquartile range. IQR is 
calulcated either within cohort or using the pooled IQR across cohorts.
The formula used is: value[subject] / (75th percentile - 25th percentile).
}
\seealso{
Other data manipulation functions: 
\code{\link{dh.dropCols}()},
\code{\link{dh.makeAgePolys}()},
\code{\link{dh.makeStrata}()}
}
\concept{data manipulation functions}
