% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-lmer-form.R
\name{dh.makeLmerForm}
\alias{dh.makeLmerForm}
\title{Make formulae for fitting multiple fractional polynomial models}
\usage{
dh.makeLmerForm(
  outcome,
  idvar,
  agevars,
  random = "intercept",
  fixed = NULL,
  age_interactions = NULL
)
}
\arguments{
\item{outcome}{outcome for the models}

\item{idvar}{unique identifier for subject}

\item{agevars}{vector of names of age polynomials in dataset}

\item{random}{either "intercept" or "slope" to specify random effects}

\item{fixed}{optional vector of fixed effects}

\item{age_interactions}{if TRUE also create interaction terms between age
terms and fixed effects#}
}
\value{
a tibble containing the created formulae
}
\description{
To identify the combination of fractional polynomials which give the best
fit often we will fit models with lots of different combinations. This
function builds formulae for that purpose.
}
