% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-lmer-form.R
\name{dh.makeLmerForm}
\alias{dh.makeLmerForm}
\title{Make formulae for fitting multiple fractional polynomial models}
\usage{
dh.makeLmerForm(
  outcome = NULL,
  id_var = NULL,
  age_vars = NULL,
  random = NULL,
  fixed = NULL,
  age_interactions = NULL
)
}
\arguments{
\item{outcome}{outcome for the models}

\item{id_var}{A character giving the name of the column within `df` which 
uniquely identifies each subject.}

\item{age_vars}{vector of names of age polynomials in dataset that you will use in your models}

\item{random}{either "intercept" or "slope" to specify random effects}

\item{fixed}{optional vector of fixed effects}

\item{age_interactions}{if TRUE also create interaction terms between age
terms and fixed effects}
}
\value{
a tibble containing the created formulae
}
\description{
This function is designed to be used with `dh.lmeMultPoly`. It generates 
formulae for multiple fractional polynomial models, which can be used as 
input to the `formula` argument in `dh.lmeMultPoly`.
}
