% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-outcome.R
\name{dh.makeOutcome}
\alias{dh.makeOutcome}
\title{Group and subset a data frame}
\usage{
dh.makeOutcome(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This was an early version of dh.makeStrata, which has in turn been deprecated in favour of
\code{dsTidyverseClient::ds.arrange() |> dsTidyverseClient::ds.group_by() |> 
dsTidyverseClient::ds.slice()}.
}
\keyword{internal}
