% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-outcome.R
\name{dh.makeOutcome}
\alias{dh.makeOutcome}
\title{This was the original version of dh.makeStrata. It has now been deprecated.}
\usage{
dh.makeOutcome(
  df = NULL,
  outcome = NULL,
  age_var = NULL,
  bands = NULL,
  mult_action = NULL,
  mult_vals = NULL,
  df_name = NULL,
  conns = NULL,
  id_var = "child_id",
  band_action = NULL,
  keep_original = NULL
)
}
\arguments{
\item{df}{See argument `df` in `dh.makeStrata`.}

\item{outcome}{Renamed to `var_to_subset` in `dh.makeStrata`}

\item{age_var}{See argument `age_var` in `dh.makeStrata`.}

\item{bands}{See argument `df` in `dh.makeStrata`}

\item{mult_action}{See argument `mult_action` in `dh.makeStrata`}

\item{mult_vals}{See argument `mult_vals` in `dh.makeStrata`}

\item{df_name}{Renamed to `new_obj` in `dh.makeStrata`}

\item{conns}{See argument `conns` in `dh.makeStrata`}

\item{id_var}{See argument `id_var` in `dh.makeStrata`}

\item{band_action}{See argument `band_action` in `dh.makeStrata`}

\item{keep_original}{Argument no longer in use in `dh.makeStrata`}
}
\description{
dh.makeStrata contains a number of improvements: it is computationally more
efficient, it is more clearly coded, and it has been more thoroughly tested.
}
