% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-outcome.R
\name{dh.makeOutcome}
\alias{dh.makeOutcome}
\title{Derives one or more outcome variable(s) from repeated measures data}
\usage{
dh.makeOutcome(
  df = NULL,
  outcome = NULL,
  age_var = NULL,
  bands = NULL,
  mult_action = c("earliest", "latest", "nearest"),
  mult_vals = NULL,
  keep_original = FALSE,
  df_name = NULL,
  conns = NULL,
  id_var = "child_id",
  band_action = c("g_l", "ge_le", "g_le", "ge_l")
)
}
\arguments{
\item{df}{opal dataframe}

\item{outcome}{name of repeated measures outcome variable}

\item{age_var}{Vector of values indicating pairs of low and high values
for which to derive outcome variables for. Must be even length}

\item{bands}{vector of alternating lower and upper age bands for variable(s)
you want to create. Variables will be derived for the age range
> lowest value and <= highest value for each band.}

\item{mult_action}{if a subject has more than one value within the time
period do we keep the earliest or latest? Default =
"earliest"}

\item{mult_vals}{if "mult_action = nearest", this argument specifies which
which value in each age band to chose values closest to
in case of multiple values}

\item{keep_original}{keep original data frame in the DataSHIELD backend}

\item{df_name}{specify data frame name on the DataSHIELD backend}

\item{conns}{connections object for DataSHIELD backends}

\item{id_var}{specify id variable (default assumes LifeCycle name 'child_id')}

\item{band_action}{specify how the values of bands are evaluated in making the subsets.
"g_l" = greater than the lowest band and less than the highest band; 
"ge_le" = greater or equal to the lowest band and less than or equal to the highest band; 
"g_le" = greater than the lowest band and less than or equal to the highest band; 
"ge_l" = greater than or equal to the lowest band and less than the highest band;}
}
\value{
a dataset containing the newly derived variables
}
\description{
Many analyses will want to use outcomes derived at a single time point,
e.g. BMI between ages 10-14. This function automates the process to do this
which is quite complex in DataSHIELD. Note that for big datasets this takes
a long time to run.
}
