% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-by-age.R
\name{dh.meanByAge}
\alias{dh.meanByAge}
\title{Calculate mean values of one repeated measures variable by an age variable}
\usage{
dh.meanByAge(
  df = NULL,
  outcome = NULL,
  age_var = NULL,
  conns = NULL,
  intervals = NULL
)
}
\arguments{
\item{df}{datashield dataframe}

\item{outcome}{outcome variable in long format}

\item{age_var}{age in years}

\item{conns}{connection object for DataSHIELD backends}

\item{intervals}{table defining our age bands}
}
\value{
Mean values for each unit of your age variable are returned
}
\description{
Calculate mean values of one repeated measures variable by an age variable
}
