% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-manual.R
\name{dh.metaManual}
\alias{dh.metaManual}
\title{Wrapper to manaully perform two-stage meta-analysis using metafor}
\usage{
dh.metaManual(model, method = "ML")
}
\arguments{
\item{model}{output from ds.glmSLMA}

\item{method}{method of meta-analyis which can be any valid method for the
metafor package}
}
\value{
list containing (i) metafor output and (ii) summary tibble of
        statistics
}
\description{
ds.SLMA functions automatically perform two-stage meta-analysis using the
metafor package. However in some circumstances you might want to use a
different package or chose different options in metaphor. This allows you
to meta-analyse multiple coefficients in one function call.
}
