% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-sep-models.R
\name{dh.metaSepModels}
\alias{dh.metaSepModels}
\title{Function in progress to meta-analyse separate models.}
\usage{
dh.metaSepModels(ref = NULL, exp = NULL, method = NULL, output = "both")
}
\arguments{
\item{ref}{Tibble, output from dh.multGlm.}

\item{exp}{Logical, whether to exponentiate coefficients after meta-analysis}

\item{method}{Method of meta-analysing coefficients.}

\item{output}{Character; "cohort" to return cohort coefficients only, "meta"
to return meta-analysed coefficients only, "both" to return both.}
}
\value{
A tibble
}
\description{
Function in progress to meta-analyse separate models.
}
\seealso{
Other model building: 
\code{\link{dh.buildModels}()}
}
\concept{model building}
