% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multGlm.R
\name{dh.multGLM}
\alias{dh.multGLM}
\title{Loop multiple GLM models and handle errors & non-convergence}
\usage{
dh.multGLM(
  df = NULL,
  ref = NULL,
  checks = TRUE,
  conns = NULL,
  vary_df = F,
  family = "gaussian"
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{ref}{reference tibble, output from dh.buildModels.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}

\item{conns}{DataSHIELD connections object.}

\item{vary_df}{Option to provide different df for different models. Default
is FALSE.}

\item{family}{Family to use in glm models. Default is "gaussian".}
}
\value{
Tibble containing five columns:
\itemize{
\item model = Description of model fit, taken from \code{model_name}.
\item formula = Formula for this model, taken from \code{formulae}.
\item cohort = Cohort for which this model was fit, taken from \code{cohort}.
\item fit = Output from glm model.
\item converged = Logical; whether or not the model converged/fit'.
}
}
\description{
This function allows you to fit different glm models to different cohorts,
for example if you want to include different covariates for different
cohorts.
}
\concept{analysis functions}
