% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-lmer.R
\name{dh.predictLmer}
\alias{dh.predictLmer}
\title{Gets predicted values based on a new dataframe for lmer models}
\usage{
dh.predictLmer(model = NULL, new_data = NULL, coh_names = NULL, newdata = NULL)
}
\arguments{
\item{model}{output model from ds.lmerSLMA function}

\item{new_data}{a dataframe or tibble with the values of the model variables
at which you want to get predicted values. All variables from the model must
be in the new data frame}

\item{coh_names}{a vector of cohort names, in the order that these were provided
in the original lmer model.}

\item{newdata}{Retired argument name. Please use `new_data' instead.}
}
\value{
A tibble of predicted outcome values based on provide variable values.
}
\description{
Currently there is no `predict` method for lmer models within DataSHIELD. 
This function replicates this, by calculating predicted values for fixed 
effects based on the model coefficients. Standard errors are returned for 
individual cohorts but yet for pooled models.
}
