% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-lmer.R
\name{dh.predictLmer}
\alias{dh.predictLmer}
\title{Gets predicted values based on a new dataframe for lmer models. Calculates
standard errors for individual cohorts but not pooled at this point.}
\usage{
dh.predictLmer(model = NULL, newdata = NULL, coh_names = NULL)
}
\arguments{
\item{model}{output model from ds.lmerSLMA function}

\item{newdata}{a dataframe or tibble with the values of the model variables
at which you want to get predicted values. All variables from the model must
be in the new data frame}

\item{coh_names}{a vector of cohort names, in the order that these were provided
in the original lmer model. #'}
}
\value{
A tibble of predicted outcome values based on provide variable values.
}
\description{
Gets predicted values based on a new dataframe for lmer models. Calculates
standard errors for individual cohorts but not pooled at this point.
}
