% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-vars.R
\name{dh.renameVars}
\alias{dh.renameVars}
\title{Rename one or more columns within a serverside data frame}
\usage{
dh.renameVars(
  df = NULL,
  current_names = NULL,
  new_names,
  conns = NULL,
  checks = TRUE
)
}
\arguments{
\item{df}{Character specifying a server-side data frame.}

\item{current_names}{Character vector of columns within `df` to rename.}

\item{new_names}{Character vector giving the new names for the columns 
specified in `current_names`.}

\item{conns}{DataSHIELD connections object.}

\item{checks}{Logical; if TRUE checks are performed prior to running the 
function. Default is TRUE.}
}
\value{
Data frame specified in `df` is returned server-side with variables
renamed.
}
\description{
This function is an analogue of `dplyr::rename` which allows you to rename 
columns of a serverside data frame.
}
\seealso{
Other data manipulation functions: 
\code{\link{dh.dropCols}()},
\code{\link{dh.makeAgePolys}()},
\code{\link{dh.makeIQR}()},
\code{\link{dh.makeStrata}()}
}
\concept{data manipulation functions}
