% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablised-weights.R
\name{dh.stablisedWeights}
\alias{dh.stablisedWeights}
\title{Generate Stabilized Weights Using ds.glmSLMA output and outcome proportions.}
\usage{
dh.stablisedWeights(
  glm_object = NULL,
  responder_df = NULL,
  responder_var = NULL,
  non_responder_level,
  responder_level,
  new_obj = NULL,
  conns = NULL
)
}
\arguments{
\item{glm_object}{A fitted ds.glmSLMA object.}

\item{responder_df}{A serverside data frame containing the responder variable.}

\item{responder_var}{Character vector of outcome variable within \code{responder_df}.}

\item{non_responder_level}{Character vector indicating level of \code{responder_var} corresponding to
non-responders.}

\item{responder_level}{Character vector indicating level of \code{responder_var} corresponding to
responders.}

\item{new_obj}{A character string for the name of the new object to store stabilized weights.}

\item{conns}{DataSHIELD connections object.}
}
\value{
Stabilized weights assigned as serverside object \code{new_obj}
}
\description{
This calculates stablised inverse probability weights. Currently only works with a binary
outcome variable.
}
\details{
The function performs the following steps:
\enumerate{
\item Generate predicted values from the GLM model.
\item Calculate proportions of a binary outcome using the outcome data.
\item Format the proportions data for analysis.
\item Create stabilized weights using the formatted proportions and GLM predictions.
}
}
\note{
Currently only works with binary outcome variable.
}
