% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subj-has-data.R
\name{dh.subjHasData}
\alias{dh.subjHasData}
\title{Identify subjects with available data}
\usage{
dh.subjHasData()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This was an early version of \code{dh.defineCases()} which has in turn been deprecated. Please
use \code{dsTidyverseClient::ds.filter()}
}
\keyword{internal}
