% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-env.R
\name{dh.tidyEnv}
\alias{dh.tidyEnv}
\title{Remove multiple objects from the serverside environment}
\usage{
dh.tidyEnv(obj = NULL, type = NULL, conns = NULL)
}
\arguments{
\item{obj}{Server-side objects that you want to either keep or remove.}

\item{type}{Either "remove" to remove objects specified in \code{obj} or "keep" to
keep objects specified in \code{obj} and remove everything else.}

\item{conns}{DataSHIELD connections object.}
}
\value{
None. Objects are removed from the server-side environnment.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This was a wrapper around ds.rm intended to allow the removal of multiple objects
in one call. It was deprecated because we realised that this was already possible with
\code{ds.rm()}
}
\keyword{internal}
