% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-env.R
\name{dh.tidyEnv}
\alias{dh.tidyEnv}
\title{Tidy up the server environment}
\usage{
dh.tidyEnv(conns = opals, obj, type = "remove", cohorts)
}
\arguments{
\item{conns}{connections object to DataSHIELD backends}

\item{obj}{objects that you want to either keep or remove}

\item{type}{either "remove" to remove the listed objects of "keep" to keep
the listed objects and remove everything else.}

\item{cohorts}{optional argument specifying which cohorts to use}
}
\value{
None. Objects removed from ds environment
}
\description{
This is a very simple wrapper around ds.rm to allow you to remove more than
one object at a time.
}
\author{
Tim Cadman
}
