% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-pred-data.R
\name{dh.trimPredData}
\alias{dh.trimPredData}
\title{Trims predicted values based on min and max values provided. Useful
where you want to ensure that predicted values are limited at the range
of observed values.}
\usage{
dh.trimPredData(
  pred = NULL,
  coh_names = NULL,
  age = "age",
  min = NULL,
  max = NULL
)
}
\arguments{
\item{pred}{a tibble of predicted values}

\item{coh_names}{a vector of cohort names corresponding to the cohort names in pred}

\item{age}{name of the variable corresponding to age in the predicted values}

\item{min}{a vector of minimum values of age, length equal to number of cohorts}

\item{max}{a vector of maximum values of age, length equal to number of cohorts}
}
\value{
The original tibble of predicted values, trimmed by the min and max ages
}
\description{
Trims predicted values based on min and max values provided. Useful
where you want to ensure that predicted values are limited at the range
of observed values.
}
