% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-pred-data.R
\name{dh.trimPredData}
\alias{dh.trimPredData}
\title{Trims predicted values based on min and max values provided}
\usage{
dh.trimPredData(
  pred = NULL,
  coh_names = NULL,
  age = "age",
  min = NULL,
  max = NULL
)
}
\arguments{
\item{pred}{a tibble of predicted values}

\item{coh_names}{a vector of cohort names corresponding to the cohort names in pred}

\item{age}{name of the var'iable corresponding to age in the predicted values}

\item{min}{a vector of minimum values of age, length equal to number of cohorts}

\item{max}{a vector of maximum values of age, length equal to number of cohorts}
}
\value{
The original tibble of predicted values, trimmed by the min and max ages
}
\description{
This can be used in conjuncture with `ds.predictLmer`. The new data frame
used for prediction may contain age or time values not present in the 
original data. This allows you to remove all predicted values beyong a min
or max age specified. This can ensure that plots only show predict values 
over a range where there was actually data.
}
