% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsHelper-defunct.R, R/make-outcome.R,
%   R/mean-by-age.R
\name{dsHelper-defunct}
\alias{dsHelper-defunct}
\alias{dh.makeOutcome}
\alias{dh.meanByAge}
\title{Defunct functions in package \pkg{dsHelper}.}
\usage{
dh.makeOutcome(...)

dh.meanByAge(...)
}
\description{
The functions listed below are defunct and no longer available.
When possible, alternative functions with similar functionality are
mentioned. Help pages for deprecated functions are available at
\code{help("<function>-defunct")}.
}
\section{\code{dh.makeOutcome}}{

For \code{dh.makeOutcome}, use \code{\link{dh.makeStrata}}.
}

\section{\code{dh.meanByAge}}{

For \code{dh.meanByAge}, use \code{\link{dh.meanByGroup}}.
}

\keyword{internal}
